clean: moved new functions to helpers
