+++ src/src/cli/tpg/commands.py
@@ -20,7 +20,7 @@ def evolve(ctx: click.Context, env: str, processes: int, seed: int):
     TPG = ctx.obj["tpg"]
     
     # Setup environment directories and get working directory
-    env_dir = create_environment_directories(TPG, env)
+    env_dir = helpers.create_environment_directories(TPG, env)
     
     # Change working directory to environment directory
     os.chdir(env_dir)
@@ -57,42 +57,6 @@ def evolve(ctx: click.Context, env: str, processes: int, seed: int):
         raise click.ClickException(f"Failed to start experiment: {e}")
 
     click.echo("Evolve (started in background)")
-    
-    
-def create_environment_directories(TPG: str, env: str):
-    """Create directory structure for environment experiments"""
-    
-    # Convert environment name to snake_case for directory naming
-    env_dir = os.path.join(TPG, "experiments", env)
-    
-    # Create directories for experiments
-    
-                # └── INSERT_ENVIRONMENT_NAME/
-                    # ├── checkpoints/
-                    # │   └── .gitignore
-                    # ├── frames/
-                    # │   └── .gitignore
-                    # └── replay/
-                    #     ├── frames/
-                    #     │   └── .gitignore
-                    #     └── graphs/
-                    #         └── .gitignore
-    directories = [
-        os.path.join(env_dir, "checkpoints"),
-        os.path.join(env_dir, "frames"),
-        os.path.join(env_dir, "replay", "frames"),
-        os.path.join(env_dir, "replay", "graphs")
-    ]
-    
-    for directory in directories:
-        os.makedirs(directory, exist_ok=True)
-        # Create .gitignore in each directory
-        gitignore_path = os.path.join(directory, ".gitignore")
-        if not os.path.exists(gitignore_path):
-            with open(gitignore_path, "w") as f:
-                f.write("*\n!.gitignore\n")
-                
-    return env_dir
 
 # TODO: Implement the plot to wrap plotting functionality
 @click.command(help="Plot results for an experiment")
@@ -114,7 +78,7 @@ def replay(ctx: click.Context, env: str, seed: int, seed_aux: int, task_to_repla
     hyper_parameters = ctx.obj["hyper_parameters"]
     TPG = ctx.obj["tpg"]
     
-    env_dir = create_environment_directories(TPG, env)
+    env_dir = helpers.create_environment_directories(TPG, env)
     
     # Change working directory to environment directory
     os.chdir(env_dir)
